refactor: TeamDetailResponse 가격관련 필드 Int -> Long으로 수정

+++ src/main/kotlin/team/mozu/dsm/adapter/in/team/dto/response/TeamDetailResponse.kt
@@ -5,10 +5,10 @@ import java.util.UUID
 data class TeamDetailResponse(
     val id: UUID,
     val teamName: String?,
-    val baseMoney: Int,
-    val totalMoney: Int,
-    val cashMoney: Int,
-    val valuationMoney: Int,
+    val baseMoney: Long,
+    val totalMoney: Long,
+    val cashMoney: Long,
+    val valuationMoney: Long,
     val curInvRound: Int,
     val valProfit: Int,
     val profitNum: Double
